loop reorder and double dtype tick

+++ common.h
@@ -7,10 +7,10 @@
 
 #define MIN(a, b) ((a) < (b) ? (a) : (b))
 
-uint64_t tick() {
+double tick() {
   struct timespec ts;
   clock_gettime(CLOCK_MONOTONIC, &ts);
-  return (uint64_t)((ts.tv_sec + ts.tv_nsec / 1e9) * 1e3);
+  return (ts.tv_sec + ts.tv_nsec / 1e9) * 1e3;
 }
 
 void print_matrix(float *m, int rows, int cols) {
